move logic to upload_custom_analysis.py
